requirements.txt
flask
flask-cors
anthropic
python-dotenv
stripe